server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via redis
bcrypt==3.2.2
    # via
    #   -r requirements.in
    #   passlib
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via
    #   bcrypt
    #   cryptography
click==8.1.7
    # via
    #   typer
    #   uvicorn
cryptography==42.0.7
    # via python-jose
dnspython==2.6.1
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.1.1
    # via fastapi
exceptiongroup==1.2.1
    # via anyio
fastapi==0.111.0
    # via
    #   -r requirements.in
    #   fastapi-cli
fastapi-camelcase==2.0.0
    # via -r requirements.in
fastapi-cli==0.0.3
    # via fastapi
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hiredis==2.3.2
    # via redis
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
mako==1.3.4
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.3
    # via fastapi
passlib[bcrypt]==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-camelcase
pydantic-core==2.18.2
    # via pydantic
pygments==2.18.0
    # via rich
pyhumps==3.8.0
    # via fastapi-camelcase
python-dotenv==1.0.1
    # via uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via uvicorn
redis[hiredis]==5.0.4
    # via -r requirements.in
rich==13.7.1
    # via typer
rsa==4.9
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.16.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.30
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via fastapi
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.11.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typer
    #   uvicorn
ujson==5.10.0
    # via fastapi
uvicorn[standard]==0.29.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
